# services/candidate-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Candidate Service",
    "description": "\n    Comprehensive candidate management and intelligent matching service.\n    \n    **Capabilities:**\n    - Candidate CRUD operations with validation\n    - Job application tracking with enum-based status\n    - Skill management with proficiency levels\n    - Resume management\n    - AI-powered candidate matching using vector search (optional)\n    - Skills-based similarity search\n    \n    **Vector Search Stack (Optional):**\n    - **FastEmbed**: ONNX-based embedding generation (no PyTorch)\n    - **LanceDB**: Embedded vector database for similarity search\n    ",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/candidates": {
      "post": {
        "tags": [
          "candidates"
        ],
        "summary": "Create a new candidate",
        "description": "Creates a new candidate record with email, name, and optional contact info",
        "operationId": "create_candidate_api_v1_candidates_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidate created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input - missing or invalid fields"
          },
          "401": {
            "description": "Unauthorized - missing authorization header"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "candidates"
        ],
        "summary": "List all candidates with pagination",
        "description": "List all candidates with pagination (offset, limit)",
        "operationId": "list_candidates_api_v1_candidates_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items to return"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of candidates",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidates/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search candidates with optional filters",
        "description": "Search for candidates with optional filters.\n\n- **query**: Search query (e.g., \"Python developer\")\n- **skills**: Comma-separated skills (e.g., \"Python,FastAPI,React\")\n- **min_experience**: Minimum years of experience\n- **location**: Location filter (e.g., \"New York\")\n- **tags**: Comma-separated tags (e.g., \"remote,full-time\")\n- **limit**: Maximum results (default: 5, max: 100)",
        "operationId": "search_candidates_api_v1_candidates_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "skills",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skills"
            }
          },
          {
            "name": "min_experience",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Experience"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidates/{candidate_id}": {
      "get": {
        "tags": [
          "candidates"
        ],
        "summary": "Get candidate by ID",
        "description": "Get candidate by ID",
        "operationId": "get_candidate_api_v1_candidates__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Candidate not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "candidates"
        ],
        "summary": "Update candidate information",
        "description": "Update candidate information",
        "operationId": "update_candidate_api_v1_candidates__candidate_id__put",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Candidate not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "candidates"
        ],
        "summary": "Delete a candidate",
        "description": "Delete a candidate",
        "operationId": "delete_candidate_api_v1_candidates__candidate_id__delete",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Candidate deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Candidate not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\+?1?\\d{9,15}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "resume_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Url"
          }
        },
        "type": "object",
        "required": [
          "email",
          "first_name",
          "last_name"
        ],
        "title": "CandidateCreate",
        "description": "Schema for creating a candidate",
        "example": {
          "email": "john.doe@example.com",
          "first_name": "John",
          "last_name": "Doe",
          "phone": "+1234567890",
          "resume_url": "https://example.com/resume.pdf"
        }
      },
...
      "CandidateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "resume_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Url"
          },
          "status": {
            "$ref": "#/components/schemas/CandidateStatus"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "phone",
          "resume_url",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "CandidateResponse",
        "description": "Schema for candidate response"
      },
      "CandidateStatus": {
        "type": "string",
        "enum": [
          "new",
          "reviewing",
          "interview_scheduled",
          "offered",
          "hired",
          "rejected"
        ],
        "title": "CandidateStatus",
        "description": "Candidate lifecycle status"
      },
...
      "CandidateUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\+?1?\\d{9,15}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "resume_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Url"
          }
        },
        "type": "object",
        "title": "CandidateUpdate",
        "description": "Schema for updating a candidate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "filters_applied": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters Applied"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results",
            "default": []
          },
          "search_method": {
            "type": "string",
            "title": "Search Method"
          }
        },
        "type": "object",
        "required": [
          "total",
          "query",
          "filters_applied",
          "search_method"
        ],
        "title": "SearchResponse"
      },
...
    }
  }
}